Refactor interactive
