+++ executorlib/__init__.py
@@ -1,12 +1,9 @@
 from typing import Optional
 
 from executorlib._version import get_versions as _get_versions
-from executorlib.interactive import create_executor
-from executorlib.interactive.dependencies import ExecutorWithDependencies
+from executorlib.interactive.executor import create_executor, ExecutorWithDependencies
 from executorlib.standalone.inputcheck import (
     check_plot_dependency_graph as _check_plot_dependency_graph,
-)
-from executorlib.standalone.inputcheck import (
     check_refresh_rate as _check_refresh_rate,
 )
